Add CommonProducer
